+++ eway/migrations/0005_auto__add_requestlog__add_ewaytransaction.py
@@ -7,6 +7,10 @@ from django.db import models
 
 class Migration(SchemaMigration):
 
+    depends_on = (
+        ("basket", "0001_initial"),
+    )
+
     def forwards(self, orm):
         # Adding model 'RequestLog'
         db.create_table('eway_requestlog', (